+++ integration-test/src/test/java/uk/gov/justice/generation/pojo/integration/examples/EnumsIT.java
@@ -2,6 +2,7 @@ package uk.gov.justice.generation.pojo.integration.examples;
 
 import static com.jayway.jsonassert.JsonAssert.with;
 import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
 import static uk.gov.justice.generation.pojo.integration.utils.PojoGeneratorPropertiesBuilder.pojoGeneratorPropertiesBuilder;
 
 import uk.gov.justice.generation.pojo.core.PojoGeneratorProperties;
@@ -9,9 +10,12 @@ import uk.gov.justice.generation.pojo.integration.utils.ClassInstantiator;
 import uk.gov.justice.generation.pojo.integration.utils.GeneratorUtil;
 import uk.gov.justice.generation.pojo.integration.utils.OutputDirectories;
 import uk.gov.justice.services.common.converter.jackson.ObjectMapperProducer;
+import uk.gov.justice.services.test.utils.core.reflection.ReflectionUtil;
 
 import java.io.File;
+import java.lang.reflect.Method;
 import java.util.List;
+import java.util.Optional;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Before;
@@ -50,13 +54,17 @@ public class EnumsIT {
                         outputDirectories);
 
         final Class<?> personWithEnumClass = classes.get(1);
-
         final Class<? extends Enum> enumClass = (Class<? extends Enum>) classes.get(0);
-        final Enum<?> red = Enum.valueOf(enumClass, "RED");
+
+        final Optional<Method> valueForMethod = ReflectionUtil.methodOf(enumClass, "valueFor");
+        assertThat(valueForMethod.isPresent(), is(true));
+
+        final Optional red = (Optional) valueForMethod.get().invoke(null, "Red");
+        assertThat(red.isPresent(), is(true));
 
         final String name = "Fred";
         final Integer age = 21;
-        final Object personWithEnum = classInstantiator.newInstance(personWithEnumClass, age, red, name);
+        final Object personWithEnum = classInstantiator.newInstance(personWithEnumClass, age, red.get(), name);
 
         final String studentJson = objectMapper.writeValueAsString(personWithEnum);