+++ integration-test/src/test/java/uk/gov/justice/generation/pojo/integration/utils/PojoGeneratorPropertiesBuilder.java
@@ -1,16 +1,13 @@
 package uk.gov.justice.generation.pojo.integration.utils;
 
-import static org.junit.Assert.assertTrue;
 import static uk.gov.justice.generation.pojo.integration.utils.TypeMappingFactory.typeMappingOf;
+import static uk.gov.justice.services.test.utils.core.reflection.ReflectionUtil.setField;
 
 import uk.gov.justice.generation.pojo.core.PojoGeneratorProperties;
 import uk.gov.justice.generation.pojo.core.TypeMapping;
 
-import java.lang.reflect.Field;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 
 public class PojoGeneratorPropertiesBuilder {
 
@@ -57,34 +54,5 @@ public class PojoGeneratorPropertiesBuilder {
         setField(pojoGeneratorProperties, "typeMappings", typeMappings);
 
         return pojoGeneratorProperties;
-    }
-
-    /**
-     * sets value of the field by reflection
-     *
-     * @param object     - object to modify
-     * @param fieldName  - name of the field belonging to the object
-     * @param fieldValue - value of the field to be set
-     * @throws IllegalAccessException if unable to access field
-     */
-    public static void setField(final Object object, final String fieldName, final Object fieldValue)
-            throws IllegalAccessException {
-        final Field field = fieldOf(object.getClass(), fieldName);
-        field.setAccessible(true);
-        field.set(object, fieldValue);
-    }
-
-    /**
-     * Searches for a field in the given class by reflection
-     *
-     * @param clazz     - class type
-     * @param fieldName - name of field in class
-     * @return - field belonging to the given clazz with the given fieldName
-     */
-    public static Field fieldOf(final Class<?> clazz, final String fieldName) {
-        final Optional<Field> field = Arrays.stream(clazz.getDeclaredFields()).filter(f -> f.getName().equals(fieldName))
-                .findFirst();
-        assertTrue(field.isPresent());
-        return field.get();
     }
 }