+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- Enum is generated with valueFor method, converts value string and returns optional enum type
+
 ## [1.0.0] - 2017-10-04
 
 ### Added